requirements.txt
pandas>=1.3.0
pyyaml>=6.0
requests>=2.26.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.910 